Khởi tạo dự án

+++ database/migrations/2024_06_30_034040_create_refresh_tokens_table.php
@@ -4,20 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('refresh_tokens', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('refresh_token');
-            $table->timestamp('expires_at')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('refresh_tokens')) {
+            Schema::create('refresh_tokens', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('refresh_token');
+                $table->timestamp('expires_at')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
